FIX/ADD OPERATIONS FUNCTIONS:
Adding all the operations provided in the subject and used in the algo
- ft_ra
- ft_rb
- ft_rr
- ft_pa
- ft_pb
- ft_sa
- ft_sb
- ft_ss

These functions are now all using the previously built functions and add
an output on stdout to explicit which one was used.
All these operations are also added in the .h file.

Swap function now uses a pointer on struct address to use the same
parameters as the other operations.