ADD FT_GET_POS.C:
Adding ft_get_pos (set the pile->pos at the correct value)
Adding ft_get_aim_a (set the pile->aim_a at the correct value) this is
the position on pile a an element of b should aim to be at.
Adding ft_get_cost (set the cost to the proper number of operations
needed for one element of pile_b to be at the position aim_a.

This file could be divided in multiple files later.