ADD FT_THE_ALGO_PART_2.c
Adding new file containing all the decision making functions.
These functions are the core of the algorithm dealing with 'same signs'
costs therefore using rr or rrr functions instead of ra/rb or rra/rrb
This is a first version that needs to be thouroughly tested, still most
of the algo is coded.