FIX FT_SORT_THREE:
Fix ft_sort_three. Ft_get_biggest was fundamentaly wrong, returning a
wrong value. Resulting in the rest of the sort not working sometimes.
Massive changes to the code structure, making it more consistant with
the rest of the work.
Function ft_sort_three should now be working as intended, sorting a pile
of 3 values.