ADD FT_GET_SMALLEST_POS && FT_GET_AIM_A_HIGHEST:
Added two new functions the former is useful to return the position of
the smallest index in one given list.
The latter is usefull to fix an incorrect value given by get_aim_a
function when the value couldn't find any higher comparison.
If this case is encountered, the function will behave as such:
the position the value will aim at is to be right before the smallest
index of the current list.