FIX MAIN:
This commit is dedicated to bsavinel the smart-ass kid who spotted this
mistake.

My main checked the args first, if only one arg was given the program
would exist doing nothing without checking if the arg given was correct
or not.

Another issue was critical, the index was set after the sort_three,
resulting in a massive issue, sort_three function dealing with unset
values.